Set piece scales to 0 when starting game(bug?)
This is to prevent flashing of completed puzzle in the beginning

I am seeing some bug where puzzle becomes in solvable state sometimes but I am not sure if it is related to this commit. (possibly yes). But it's hard to reproduce the bug and I don't know why it happens.